requirements.txt
faiss-cpu==1.13.2
numpy>=1.24,<2.0
opencv-python
insightface==0.7.3
onnxruntime==1.23.2